+++ src/tmc2209_driver.cpp
@@ -139,7 +139,7 @@ void Tmc2209Driver::setStepState(bool state)
   stepper_common::stepper_write_gpio(STEP_OUTPUT_PORT, STEP_OUT_PIN, state);
 }
 
-void Tmc2209Driver::sendStepPulse(uint32_t width_ns)
+void Tmc2209Driver::sendStepPulse(uint64_t width_ns)
 {
   stepper_common::stepper_write_gpio_high(STEP_OUTPUT_PORT, STEP_OUT_PIN);
   delay_ns(width_ns);
